make gdt with C
