+++ src/string.c
@@ -1,0 +1,11 @@
+#include <string.h>
+#include <stdint.h>
+#include <stddef.h>
+int strlen(char* str) {
+    if(str == NULL)
+        return 0;
+
+    size_t cnt = 0;
+    while(str[cnt]) cnt++;
+    return cnt;
+}
