+++ src/string.h
@@ -1,0 +1,6 @@
+#ifndef STRING_H_INCLUDED
+#define STRING_H_INCLUDED
+
+int strlen(char* str);
+
+#endif // STRING_H_INCLUDED
